feat(STT): added SpeechToText module
